ninth commit

+++ proj2/gitlet/Checkout.java
@@ -23,21 +23,26 @@ public class Checkout {
     private void checkoutOverride1(String Branch) throws IOException, ClassNotFoundException {
 
         String lastCommitHashCode = Utils.readContentsAsString(new File(Main.Branches, Branch));
-        File lastCommitFile = new File(Main.Objects,lastCommitHashCode);
-        ObjectInputStream ois = new ObjectInputStream(new FileInputStream(lastCommitFile));
-        Commit lastCommit = (Commit) ois.readObject();
-        ois.close();
+        if (lastCommitHashCode == null || lastCommitHashCode.length() == 0) {
 
-        for (File cwdFile : Main.CWD.listFiles()){
-            cwdFile.delete();
         }
+        else {
+            File lastCommitFile = new File(Main.Objects,lastCommitHashCode);
+            ObjectInputStream ois = new ObjectInputStream(new FileInputStream(lastCommitFile));
+            Commit lastCommit = (Commit) ois.readObject();
+            ois.close();
 
-        for (String file : lastCommit.HashMapBlobs.keySet()) {
-            HashMap<String, Blobs> HashBlobs = lastCommit.HashMapBlobs.get(file);
-            File writeFile = new File(Main.CWD,file);
-            clearFile(writeFile);
-            for (Blobs blobs : HashBlobs.values()) {
-                Utils.writeContents(writeFile,blobs.getContent());
+            for (File cwdFile : Main.CWD.listFiles()){
+                cwdFile.delete();
+            }
+
+            for (String file : lastCommit.HashMapBlobs.keySet()) {
+                HashMap<String, Blobs> HashBlobs = lastCommit.HashMapBlobs.get(file);
+                File writeFile = new File(Main.CWD,file);
+                clearFile(writeFile);
+                for (Blobs blobs : HashBlobs.values()) {
+                    Utils.writeContents(writeFile,blobs.getContent());
+                }
             }
         }
     }
